[settings] remove `relativenumber`
This improves performance and also helps me to use movements that are
stored in the jumplist `[n]G` instead of `[count]j` or `[count]k` which
ins't stored in the jumplist.

I'm starting to rely in the jumplist so I need vertical mouse movements that are compatible with that funcionality.