[telescope] understand ripgrep, fd, fzf in telescope
Asumming you have installed:

- ripgrep
- fd
- fzf
- telescope-fzf-native.nvim

Almost all your telescope searches prioritize `ripgrep` as the grep
tool, in the case of `live_grep` it is required. In the case of
`find_files` I believe it uses `fd` instead of ripgrep.

What about fzf? It is only used as the sorter. Sorting can only happen
**after** you had received the search results from `ripgrep`. In other words,
once telescope presents to you the initial grep search results, if you want to further
sort/filter these results, you will have to manually do an additional
search and this is the fuzzy search.

This is the default behaviour for `grep_string` which finds the word
under the cursor, but for `live_grep` you'll have to press `C-Space` to
be able to refine your initial results with a fuzzy search.